backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
sqlalchemy==2.0.23
pymysql==1.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
openai==1.6.1
